Handle all file read operations in files.ts.
Add support for future file formats/format versioning
